+++ src/python/deepgnn/graph_engine/snark/dispatcher.py
@@ -60,7 +60,7 @@ class PipeDispatcher(Dispatcher):
         folder: str,
         parallel: int,
         meta: str,
-        decoder_class: Optional[DecoderType],
+        decoder: Optional[DecoderType],
         process: typing.Callable[
             [
                 typing.Union[mp.Queue, Connection],
@@ -87,7 +87,7 @@ class PipeDispatcher(Dispatcher):
             parallel (int): Number of parallel process to use for conversion.
             meta (str): Meta data about graph.
             process (typing.Callable[ [typing.Union[mp.Queue, Connection], mp.Queue, str, int, int, int, Decoder], None ]): Function to call for processing lines in a file.
-            decoder_class (Decoder): Decoder object which is used to parse the raw graph data file.
+            decoder (Decoder): Decoder object which is used to parse the raw graph data file.
             partition_offset(int): offset in a text file, where to start reading for a new partition.
             use_threads(bool): use threads instead of processes for parallel processing.
             skip_node_sampler(bool): skip generation of node alias tables.
@@ -115,7 +115,7 @@ class PipeDispatcher(Dispatcher):
                     i + partition_offset,
                     int(self.jsm["node_type_num"]),
                     int(self.jsm["edge_type_num"]),
-                    decoder_class,
+                    decoder,
                     skip_node_sampler,
                     skip_edge_sampler,
                 ),
@@ -190,7 +190,7 @@ class QueueDispatcher(Dispatcher):
         num_partitions: int,
         meta: str,
         partion_func: typing.Callable[[str], int],
-        decoder_class: Optional[DecoderType],
+        decoder: Optional[DecoderType],
         process: typing.Callable[
             [
                 typing.Union[mp.Queue, Connection],
@@ -218,7 +218,7 @@ class QueueDispatcher(Dispatcher):
             meta (str): meta data about graph.
             process (typing.Callable[[mp.Queue, mp.Queue, str, int, int], None]): function to use for conversion.
             partion_func (typing.Callable[[str], int]): how to assign graph elements to a partition.
-            decoder_class (Decoder): Decoder object which is used to parse the raw graph data file.
+            decoder (Decoder): Decoder object which is used to parse the raw graph data file.
             partition_offset(int): offset in a text file, where to start reading for a new partition.
             use_threads(bool): use threads instead of processes for parallel processing.
             skip_node_sampler(bool): skip generation of node alias tables.
@@ -250,7 +250,7 @@ class QueueDispatcher(Dispatcher):
                     i + partition_offset,
                     int(self.jsm["node_type_num"]),
                     int(self.jsm["edge_type_num"]),
-                    decoder_class,
+                    decoder,
                     skip_node_sampler,
                     skip_edge_sampler,
                 ),